// frontend/types.ts
export interface CareerStep {
    title: string;
    description: string;
    skills_to_acquire: string[];
}

export interface CareerPath {
    path_name: string;
    steps: CareerStep[];
}